+++ regression/reference/example/utilUserLibrary.cpp
@@ -7,6 +7,9 @@
 // SPDX-License-Identifier: (BSD-3-Clause)
 //
 
+// typemap
+#include "ExClass1.hpp"
+#include "ExClass2.hpp"
 // shroud
 #include "typesUserLibrary.h"
 #include <cstddef>
@@ -28,6 +31,40 @@ void AA_ShroudCopyStringAndFree(AA_SHROUD_array *data, char *c_var, size_t c_var
     AA_SHROUD_memory_destructor(&data->cxx); // delete data->cxx.addr
 }
 
+
+// Release library allocated memory.
+void AA_SHROUD_memory_destructor(AA_SHROUD_capsule_data *cap)
+{
+    void *ptr = cap->addr;
+    switch (cap->idtor) {
+    case 0:   // --none--
+    {
+        // Nothing to delete
+        break;
+    }
+    case 1:   // example::nested::ExClass1
+    {
+        example::nested::ExClass1 *cxx_ptr = 
+            reinterpret_cast<example::nested::ExClass1 *>(ptr);
+        delete cxx_ptr;
+        break;
+    }
+    case 2:   // example::nested::ExClass2
+    {
+        example::nested::ExClass2 *cxx_ptr = 
+            reinterpret_cast<example::nested::ExClass2 *>(ptr);
+        delete cxx_ptr;
+        break;
+    }
+    default:
+    {
+        // Unexpected case in destructor
+        break;
+    }
+    }
+    cap->addr = nullptr;
+    cap->idtor = 0;  // avoid deleting again
+}
 
 #ifdef __cplusplus
 }