+++ regression/reference/example/wrapUserLibrary.cpp
@@ -7,13 +7,6 @@
 // SPDX-License-Identifier: (BSD-3-Clause)
 //
 
-// typemap
-#include "ExClass1.hpp"
-#include "ExClass2.hpp"
-// shroud
-#include "typesUserLibrary.h"
-#include <cstdlib>
-
 // splicer begin CXX_definitions
 //   CXX_definitions
 // splicer end CXX_definitions
@@ -22,39 +15,5 @@ extern "C" {
 
 // splicer begin C_definitions
 // splicer end C_definitions
-
-// Release library allocated memory.
-void AA_SHROUD_memory_destructor(AA_SHROUD_capsule_data *cap)
-{
-    void *ptr = cap->addr;
-    switch (cap->idtor) {
-    case 0:   // --none--
-    {
-        // Nothing to delete
-        break;
-    }
-    case 1:   // example::nested::ExClass1
-    {
-        example::nested::ExClass1 *cxx_ptr = 
-            reinterpret_cast<example::nested::ExClass1 *>(ptr);
-        delete cxx_ptr;
-        break;
-    }
-    case 2:   // example::nested::ExClass2
-    {
-        example::nested::ExClass2 *cxx_ptr = 
-            reinterpret_cast<example::nested::ExClass2 *>(ptr);
-        delete cxx_ptr;
-        break;
-    }
-    default:
-    {
-        // Unexpected case in destructor
-        break;
-    }
-    }
-    cap->addr = nullptr;
-    cap->idtor = 0;  // avoid deleting again
-}
 
 }  // extern "C"