+++ regression/reference/names/top.cpp
@@ -12,7 +12,6 @@
 #include <string>
 // shroud
 #include <cstring>
-#include "typestestnames.hh"
 #include <cstdlib>
 
 // splicer begin CXX_definitions
@@ -348,32 +347,6 @@ int TES_cstruct_as_class_sum(TES_Cstruct_as_class * point)
     int SHC_rv = Cstruct_as_class_sum(ARG_point);
     return SHC_rv;
     // splicer end function.cstruct_as_class_sum
-}
-
-// Release library allocated memory.
-void TES_SHROUD_memory_destructor(TES_SHROUD_capsule_data *cap)
-{
-    void *ptr = cap->addr;
-    switch (cap->idtor) {
-    case 0:   // --none--
-    {
-        // Nothing to delete
-        break;
-    }
-    case 1:   // ns0::Names
-    {
-        ns0::Names *cxx_ptr = reinterpret_cast<ns0::Names *>(ptr);
-        delete cxx_ptr;
-        break;
-    }
-    default:
-    {
-        // Unexpected case in destructor
-        break;
-    }
-    }
-    cap->addr = nullptr;
-    cap->idtor = 0;  // avoid deleting again
 }
 
 }  // extern "C"