+++ regression/reference/namespace/utilns.cpp
@@ -29,6 +29,13 @@ void NS_ShroudCopyStringAndFree(NS_SHROUD_array *data, char *c_var, size_t c_var
 }
 
 
+// Release library allocated memory.
+void NS_SHROUD_memory_destructor(NS_SHROUD_capsule_data *cap)
+{
+    cap->addr = nullptr;
+    cap->idtor = 0;  // avoid deleting again
+}
+
 #ifdef __cplusplus
 }
 #endif