+++ regression/reference/strings-cfi/wrapstrings.cpp
@@ -14,7 +14,6 @@
 #include <string>
 // shroud
 #include <cstring>
-#include "typesstrings.h"
 #include <cstdlib>
 
 // splicer begin CXX_definitions
@@ -1493,47 +1492,5 @@ void STR_post_declare_CFI(int * count, CFI_cdesc_t *SHcfi_name)
         SHCXX_name.size());
     // splicer end function.post_declare_CFI
 }
-
-// start release allocated memory
-// Release library allocated memory.
-void STR_SHROUD_memory_destructor(STR_SHROUD_capsule_data *cap)
-{
-    void *ptr = cap->addr;
-    switch (cap->idtor) {
-    case 0:   // --none--
-    {
-        // Nothing to delete
-        break;
-    }
-    case 1:   // new_string
-    {
-        std::string *cxx_ptr = reinterpret_cast<std::string *>(ptr);
-        delete cxx_ptr;
-        break;
-    }
-    case 2:   // std::string
-    {
-        std::string *cxx_ptr = reinterpret_cast<std::string *>(ptr);
-        delete cxx_ptr;
-        break;
-    }
-    case 3:   // C_string_free
-    {
-        // Used with +free_pattern(C_string_free)
-        std::string *cxx_ptr = reinterpret_cast<std::string *>(ptr);
-        delete cxx_ptr;
-
-        break;
-    }
-    default:
-    {
-        // Unexpected case in destructor
-        break;
-    }
-    }
-    cap->addr = nullptr;
-    cap->idtor = 0;  // avoid deleting again
-}
-// end release allocated memory
 
 }  // extern "C"