+++ regression/reference/tutorial/utilTutorial.cpp
@@ -7,6 +7,8 @@
 // SPDX-License-Identifier: (BSD-3-Clause)
 //
 
+// typemap
+#include <string>
 // shroud
 #include "typesTutorial.h"
 #include <cstddef>
@@ -30,6 +32,34 @@ void TUT_ShroudCopyStringAndFree(TUT_SHROUD_array *data, char *c_var, size_t c_v
 }
 // end helper copy_string
 
+
+// start release allocated memory
+// Release library allocated memory.
+void TUT_SHROUD_memory_destructor(TUT_SHROUD_capsule_data *cap)
+{
+    void *ptr = cap->addr;
+    switch (cap->idtor) {
+    case 0:   // --none--
+    {
+        // Nothing to delete
+        break;
+    }
+    case 1:   // new_string
+    {
+        std::string *cxx_ptr = reinterpret_cast<std::string *>(ptr);
+        delete cxx_ptr;
+        break;
+    }
+    default:
+    {
+        // Unexpected case in destructor
+        break;
+    }
+    }
+    cap->addr = nullptr;
+    cap->idtor = 0;  // avoid deleting again
+}
+// end release allocated memory
 
 #ifdef __cplusplus
 }