+++ regression/reference/vectors/utilvectors.cpp
@@ -7,6 +7,8 @@
 // SPDX-License-Identifier: (BSD-3-Clause)
 //
 
+// typemap
+#include <vector>
 // shroud
 #include "typesvectors.h"
 #include <cstring>
@@ -31,6 +33,42 @@ void VEC_ShroudCopyArray(VEC_SHROUD_array *data, void *c_var,
     VEC_SHROUD_memory_destructor(&data->cxx); // delete data->cxx.addr
 }
 // end helper copy_array
+
+// start release allocated memory
+// Release library allocated memory.
+void VEC_SHROUD_memory_destructor(VEC_SHROUD_capsule_data *cap)
+{
+    void *ptr = cap->addr;
+    switch (cap->idtor) {
+    case 0:   // --none--
+    {
+        // Nothing to delete
+        break;
+    }
+    case 1:   // std_vector_int
+    {
+        std::vector<int> *cxx_ptr = 
+            reinterpret_cast<std::vector<int> *>(ptr);
+        delete cxx_ptr;
+        break;
+    }
+    case 2:   // std_vector_double
+    {
+        std::vector<double> *cxx_ptr = 
+            reinterpret_cast<std::vector<double> *>(ptr);
+        delete cxx_ptr;
+        break;
+    }
+    default:
+    {
+        // Unexpected case in destructor
+        break;
+    }
+    }
+    cap->addr = nullptr;
+    cap->idtor = 0;  // avoid deleting again
+}
+// end release allocated memory
 
 #ifdef __cplusplus
 }