+++ CHANGELOG.md
@@ -46,6 +46,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   fixes a compile error with xlf.
 
 ### Changed
+- The *C_memory_dtor_function* is now written to the utility file,
+  *C_impl_utility*.  This function contains code to delete memory from
+  shadow classes. Previously it was written file *C_impl_filename*.
+  This may require changes to Makefiles.
 
 ### Removed
 - Removed format field *F_capsule_data_type_class*.